implement revised symbol analysis pass
