+++ lace/include/lace/tree/Codegen.h
@@ -1,0 +1,19 @@
+//
+//  Copyright (c) 2025-2026 Nicholas Marino
+//  All rights reserved.
+//
+
+#ifndef LACE_CODEGEN_H_
+#define LACE_CODEGEN_H_
+
+#include "lace/tree/VisitorBase.h"
+
+namespace lace {
+
+class Codegen final : public VisitorBase {
+
+};
+
+} // namespace lace
+
+#endif // LACE_CODEGEN_H_
