+++ lace/source/tree/VisitorBase.cpp
@@ -3,7 +3,7 @@
 //  All rights reserved.
 //
 
-#include "lace/tree/AST.h"
+#include "lace/tree/Rib.h"
 #include "lace/tree/Defn.h"
 #include "lace/tree/Expr.h"
 #include "lace/tree/Stmt.h"
@@ -11,11 +11,10 @@
 
 using namespace lace;
 
-void VisitorBase::visit(AST& node) {
-    m_ast = &node;
-    m_scope = node.scope();
+void VisitorBase::visit(Rib& rib) {
+    m_rib = &rib;
 
-    for (Defn* defn : node.defns())
+    for (Defn* defn : rib.defns())
         defn->accept(*this);
 }
 
@@ -33,8 +32,6 @@ void VisitorBase::visit(FieldDefn& node) {
 }
 
 void VisitorBase::visit(FunctionDefn& node) {
-    m_scope = node.scope();
-
     if (node.has_receiver())
         node.receiver()->accept(*this);
 
@@ -43,30 +40,10 @@ void VisitorBase::visit(FunctionDefn& node) {
 
     if (node.has_body())
         node.body()->accept(*this);
-
-    m_scope = m_scope->parent();
-}
-
-void VisitorBase::visit(LoadDefn& node) {
-
 }
 
 void VisitorBase::visit(ParameterDefn& node) {
 
-}
-
-void VisitorBase::visit(SpaceDefn& node) {
-    m_namespaces.push_back(&node);
-    m_scope = node.scope();
-
-    for (NamedDefn* defn : node.defns()) {
-        // Only pass over definitions defined in the same file.
-        if (defn->origin() == node.origin())
-            defn->accept(*this);
-    }
-
-    m_scope = m_scope->parent();
-    m_namespaces.pop_back();
 }
 
 void VisitorBase::visit(StructDefn& node) {
@@ -96,12 +73,8 @@ void VisitorBase::visit(AdapterStmt& node) {
 }
 
 void VisitorBase::visit(BlockStmt& node) {
-    m_scope = node.scope();
-
     for (Stmt* stmt : node.stmts())
         stmt->accept(*this);
-
-    m_scope = m_scope->parent();
 }
 
 void VisitorBase::visit(IfStmt& node) {
@@ -204,52 +177,4 @@ void VisitorBase::visit(StructInitExpr& node) {
 void VisitorBase::visit(SubscriptExpr& node) {
     node.base()->accept(*this);
     node.index()->accept(*this);
-}
-
-Type* VisitorBase::resolve_type(Type* type) const {
-    if (auto deferred = dynamic_cast<DeferredType*>(type)) {
-        NamedDefn* named_defn = m_scope->get(deferred->name()); 
-        if (!named_defn)
-            return nullptr;
-
-        TypeDefn* type_defn = dynamic_cast<TypeDefn*>(named_defn);
-        if (!type_defn)
-            return nullptr;
-
-        return type_defn->type();
-    } else if (auto enumeration = dynamic_cast<EnumType*>(type)) {
-        Type* underlying = resolve_type(enumeration->underlying());
-        if (underlying != enumeration->underlying())
-            enumeration->set_underlying(underlying);
-
-        return enumeration;
-    } else if (auto func = dynamic_cast<FunctionType*>(type)) {
-        Type* result = resolve_type(func->result());
-        if (!result)
-            return nullptr;
-
-        std::vector<Type*> params = {};
-        params.reserve(func->num_params());
-
-        for (Type* param : func->params()) {
-            Type* res = resolve_type(param);
-            if (!res)
-                return nullptr;
-
-            params.push_back(res);
-        }
-
-        return FunctionType::get(*m_ast, result, params);
-    } else if (auto ptr = dynamic_cast<PointerType*>(type)) {
-        Type* pointee = resolve_type(ptr->pointee());
-        if (!pointee)
-            return nullptr;
-
-        if (pointee != ptr->pointee())
-            ptr->set_pointee(pointee);
-
-        return ptr;
-    }
-
-    return type;
 }
